Create netutil
