+++ core/miner/miner_test.go
@@ -1,4 +1,4 @@
-package test
+package miner 
 
 /*
 import (
